+++ src/parallel.cpp
@@ -1,4 +1,5 @@
 // [[Rcpp::depends(RcppArmadillo)]]
+#define ARMA_DONT_PRINT_ERRORS
 #include <RcppArmadillo.h>
 
 using namespace Rcpp;
@@ -13,9 +14,10 @@ using namespace arma;
 //' @param n_vars numeric. Number of variables / indicators in dataset.
 //' @param N numeric. Number of cases / observations in dataset.
 //' @param eigen_type numeric. Whether PCA (eigen_type = 1; i.e., leaving diagonal of correlation matrix at 1) or PAF (eigen_type = 2; i.e., setting diagonal of correlation matrix to SMCs).
+//' @param maxit numeric. Maximum iterations to perform after which to abort.
 // [[Rcpp::export(.parallel_sim)]]
 arma::mat parallel_sim(const int n_datasets, const int n_vars, const int N,
-                         const int eigen_type) {
+                         const int eigen_type, const int maxit = 10000) {
   // initialize needed objects
   arma::vec Lambda(n_vars);
   arma::vec eigval(n_vars);
@@ -37,16 +39,38 @@ arma::mat parallel_sim(const int n_datasets, const int n_vars, const int N,
   } else if (eigen_type == 2) { // SMC
     arma::vec smc(n_vars);
     arma::mat temp(n_vars, n_vars);
+    int success = 0;
+    int iter = 0;
 
-    for (uword i = 0; i < n_datasets; i++) {
+    while((success < n_datasets) && (iter < maxit)) {
       x = randn(N, n_vars);
       R = cor(x);
-      temp = inv_sympd(R);
+      bool flag = inv_sympd(temp, R);
+      if (!flag){
+        iter++;
+        continue;
+      }
       R.diag() = 1 - (1 / temp.diag());
       eig_sym(eigval, R);
       Lambda = flipud(eigval);
-      eig_vals.row(i) = Lambda.t();
+      eig_vals.row(success) = Lambda.t();
+      iter++;
+      success++;
     }
+
+    if ((iter == maxit) && (success < n_datasets)) {
+      stop("Could not generate enough non-singular matrices.");
+    }
+
+    // for (uword i = 0; i < n_datasets; i++) {
+    //   x = randn(N, n_vars);
+    //   R = cor(x);
+    //   temp = inv(R); // previously inv_sympd()
+    //   R.diag() = 1 - (1 / temp.diag());
+    //   eig_sym(eigval, R);
+    //   Lambda = flipud(eigval);
+    //   eig_vals.row(i) = Lambda.t();
+    // }
 
   }